+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/implementation/EmbeddingsUtils.java
@@ -3,13 +3,19 @@
 
 package com.azure.ai.openai.implementation;
 
+import com.azure.core.util.Base64Util;
 import com.azure.core.util.BinaryData;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.nio.FloatBuffer;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.List;
 
 /** This class contains convenience methods and constants for operations related to Embeddings */
 public final class EmbeddingsUtils {
@@ -25,5 +31,18 @@ public final class EmbeddingsUtils {
         }
 
         return inputJson;
+    }
+
+    // This method converts a base64 string to a list of floats
+    public static List<Float> convertBase64ToFloatList(String embedding) {
+        byte[] bytes = Base64Util.decodeString(embedding);
+        ByteBuffer byteBuffer = ByteBuffer.wrap(bytes);
+        byteBuffer.order(ByteOrder.LITTLE_ENDIAN);
+        FloatBuffer floatBuffer = byteBuffer.asFloatBuffer();
+        List<Float> floatList = new ArrayList<>(floatBuffer.remaining());
+        while (floatBuffer.hasRemaining()) {
+            floatList.add(floatBuffer.get());
+        }
+        return floatList;
     }
 }